Inception
This is my first commit to this project. I have set up secure login for
the admin using jee6's security features. I have allowed for the
connection with the database for the creation of masters.
side note : every created master will be added to an xhtml file in order
to lighten the charge on the database.
I am still learning jsf , many errors may be hidden within the code.