It basically works
So now it basically works. I changed the one page with two tabs to 3
pages. Initially I had a problem reloading the xhtml file from the
second tab to make the changes appear in the first tab. Then I found out
that I can't include files with special characters like é or è etc... So
for now in my examples I am not using any of those. Of course I will
have to find a solution to that.